requirements.txt
python-dotenv
lyricsgenius
yandex-music
